+++ src/AES256CTR.ts
@@ -1,4 +1,4 @@
-import Cipher from '~/Cipher';
+import Cipher from './Cipher';
 
 /**
  * AES-256-CTR cipher.