+++ src/index.ts
@@ -2,15 +2,15 @@
  * cryptographic-js — Lightweight AES encryption and decryption for Node.js.
  * @module cryptographic-js
  */
-import AES256CTR from '~/AES256CTR';
-import AES192CTR from '~/AES192CTR';
-import AES128CTR from '~/AES128CTR';
-import AES256CBC from '~/AES256CBC';
-import AES192CBC from '~/AES192CBC';
-import AES128CBC from '~/AES128CBC';
-import AES256CFB from '~/AES256CFB';
-import AES192CFB from '~/AES192CFB';
-import AES128CFB from '~/AES128CFB';
+import AES256CTR from './AES256CTR';
+import AES192CTR from './AES192CTR';
+import AES128CTR from './AES128CTR';
+import AES256CBC from './AES256CBC';
+import AES192CBC from './AES192CBC';
+import AES128CBC from './AES128CBC';
+import AES256CFB from './AES256CFB';
+import AES192CFB from './AES192CFB';
+import AES128CFB from './AES128CFB';
 
 export {
   AES256CTR,