+++ types/AES128CBC.d.ts
@@ -1,4 +1,4 @@
-import Cipher from '~/Cipher';
+import Cipher from './Cipher';
 /**
  * AES-128-CBC cipher.
  * 128-bit key, 16-byte IV, Cipher Block Chaining mode.