+++ types/index.d.ts
@@ -2,13 +2,13 @@
  * cryptographic-js — Lightweight AES encryption and decryption for Node.js.
  * @module cryptographic-js
  */
-import AES256CTR from '~/AES256CTR';
-import AES192CTR from '~/AES192CTR';
-import AES128CTR from '~/AES128CTR';
-import AES256CBC from '~/AES256CBC';
-import AES192CBC from '~/AES192CBC';
-import AES128CBC from '~/AES128CBC';
-import AES256CFB from '~/AES256CFB';
-import AES192CFB from '~/AES192CFB';
-import AES128CFB from '~/AES128CFB';
+import AES256CTR from './AES256CTR';
+import AES192CTR from './AES192CTR';
+import AES128CTR from './AES128CTR';
+import AES256CBC from './AES256CBC';
+import AES192CBC from './AES192CBC';
+import AES128CBC from './AES128CBC';
+import AES256CFB from './AES256CFB';
+import AES192CFB from './AES192CFB';
+import AES128CFB from './AES128CFB';
 export { AES256CTR, AES192CTR, AES128CTR, AES256CBC, AES192CBC, AES128CBC, AES256CFB, AES192CFB, AES128CFB };